requirements.txt
Django>=3.0.6,<=4.0.10
uWSGI>=2.0.18,<2.1
asgiref>=3.2.10
django-filter==2.4.0
django-cors-headers==3.5.0
djangorestframework==3.12.1
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
django-debug-toolbar==3.1.1
djangorestframework-simplejwt==4.6.0
phonenumbers==8.12.18
django-crontab==0.7.1
requests==2.25.1

